requirements.txt
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.4
openai==1.58.1
pinecone-client==5.0.1
selenium==4.27.1
streamlit==1.40.2
tiktoken==0.8.0
chromadb==0.5.23
python-dotenv==1.0.1
faiss-cpu==1.9.0
numexpr==2.10.2